backend/requirements_additions.txt
# Additional requirements for enhanced MAESTRO backend

# LangChain & LangGraph
langchain>=0.1.0
langgraph>=0.0.20
langchain-community>=0.0.20
langchain-google-genai>=0.0.5

# Vector Stores & Embeddings
chromadb>=0.4.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# Database & Caching
redis>=5.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Async & Task Queue
celery>=5.3.0
aioredis>=2.0.0
asyncio>=3.4.3

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0
pyyaml>=6.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# API & Web
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Monitoring & Logging
prometheus-client>=0.19.0
structlog>=23.2.0

# Utilities
python-dotenv>=1.0.0
tenacity>=8.2.0  # For retry logic
